edit skeleton and not found page
