+++ src/layout/SideNav.jsx
@@ -6,15 +6,26 @@ import Button from "../components/Button";
 import Search from "../features/filter/Search";
 import Categories from "../features/filter/Categories";
 import CategorieItem from "../features/filter/CategorieItem";
+import { useEffect, useState } from "react";
+import Skeleton from "react-loading-skeleton";
 
 function SideNav() {
     const { pathname } = useLocation();
+    const [skeleton, setSkeleton] = useState(false);
+
+    useEffect(() => {
+        const id = setTimeout(() => {
+            setSkeleton(true);
+        }, 1200);
+        return () => clearTimeout(id);
+    }, []);
 
     return (
         <>
             <div className="flex flex-col gap-2 pb-4 border-b-2">
                 <div className="sticky">
-                    <Search div={"pt-10 pb-2"} />
+                    <div className="text-2xl font-semibold pl-5 mt-9">Marketplace</div>
+                    <Search nameTagSearch="" div={" "} />
                     <div className="px-4  ">
                         <CategorieItem icons={<FaStore />} title={"Browse All"} isActive={pathname === "/"} />
                     </div>
@@ -37,7 +48,7 @@ function SideNav() {
                 </div>
                 <div className="flex justify-center w-full">
                     <Link to={"/create"}>
-                        <Button text={"Create new listing"} />
+                        {skeleton ? <Button text={"Create new listing"} /> : <Skeleton width={200} height={45} />}
                     </Link>
                 </div>
             </div>
