+++ src/layout/SideNavCategory.jsx
@@ -10,6 +10,7 @@ import { useEffect, useState } from "react";
 import validateSchema from "../utils/validate-schema";
 import { filterPriceSchema } from "../utils/product-validator";
 import InputErrorMessage from "../features/auth/InputErrorMessage";
+import Skeleton from "react-loading-skeleton";
 
 function SideNavCategory() {
     const dispatch = useDispatch();
@@ -17,6 +18,14 @@ function SideNavCategory() {
     const { productPrice } = useSelector((state) => state.product);
     const nameTagSearch = pathname.slice(10).replace(/_/g, " ");
     const [error, setError] = useState({});
+    const [skeleton, setSkeleton] = useState(false);
+
+    useEffect(() => {
+        const id = setTimeout(() => {
+            setSkeleton(true);
+        }, 1200);
+        return () => clearTimeout(id);
+    }, []);
 
     useEffect(() => {
         setError({});
@@ -52,41 +61,47 @@ function SideNavCategory() {
                 </div>
                 <div className="flex justify-center w-full">
                     <Link className="w-full px-20" to={"/create"}>
-                        <Button text={"Create new listing"} />
+                        {skeleton ? <Button text={"Create new listing"} /> : <Skeleton width={200} height={45} />}
                     </Link>
                 </div>
             </div>
 
             <div className="flex flex-col p-4">
                 <div className="text-lg font-semibold">Filters</div>
-                <div className="flex justify-between items-center px-4">
-                    <div className="w-32">
-                        <InputForm
-                            name="minPrice"
-                            placeholder="Min"
-                            className="pt-0"
-                            onChange={onChangeInput}
-                            value={productPrice.minPrice}
-                            isError={error.productPrice}
-                        />
-                        {error.minPrice && <InputErrorMessage message={"Minimum price must be a number"} />}
-                    </div>
+                {skeleton ? (
+                    <div className="flex justify-between items-center px-4">
+                        <div className="w-32">
+                            <InputForm
+                                name="minPrice"
+                                placeholder="Min"
+                                className="pt-0"
+                                onChange={onChangeInput}
+                                value={productPrice.minPrice}
+                                isError={error.productPrice}
+                            />
 
-                    <div>
-                        <div>To</div>
+                            {error.minPrice && <InputErrorMessage message={"Minimum price must be a number"} />}
+                        </div>
+
+                        <div>
+                            <div>To</div>
+                        </div>
+                        <div className="w-32">
+                            <InputForm
+                                name="maxPrice"
+                                placeholder="Max"
+                                className="pt-0"
+                                onChange={onChangeInput}
+                                value={productPrice.maxPrice}
+                                isError={error.productPrice}
+                            />
+
+                            {error.maxPrice && <InputErrorMessage message={"Maximum price must be a number"} />}
+                        </div>
                     </div>
-                    <div className="w-32">
-                        <InputForm
-                            name="maxPrice"
-                            placeholder="Max"
-                            className="pt-0"
-                            onChange={onChangeInput}
-                            value={productPrice.maxPrice}
-                            isError={error.productPrice}
-                        />
-                        {error.maxPrice && <InputErrorMessage message={"Maximum price must be a number"} />}
-                    </div>
-                </div>
+                ) : (
+                    <Skeleton width={320} height={50} />
+                )}
             </div>
 
             <div className="border" />
