+++ src/pages/CreateProductPage.jsx
@@ -7,9 +7,11 @@ import { fetchAllProduct, fetchProductByUserId } from "../stores/slices/productS
 import item from "../assets/Images/item.png";
 import vehicle from "../assets/Images/vehicle.png";
 import home from "../assets/Images/home.png";
+import Skeleton from "react-loading-skeleton";
 
 function CreateProductPage() {
     const dispatch = useDispatch();
+    const [skeleton, setSkeleton] = useState(false);
 
     useEffect(() => {
         const id = setTimeout(() => {
@@ -71,7 +73,11 @@ function CreateProductPage() {
                 <div className="min-w-[300px]"></div>
                 <div className="flex justify-center items-center w-full">
                     <div className="flex gap-3 flex-col">
-                        <div className="text-xl font-bold">Choose Listing Type</div>
+                        {skeleton ? (
+                            <div className="text-xl font-bold">Choose Listing Type</div>
+                        ) : (
+                            <Skeleton width={180} />
+                        )}
 
                         <div className="grid grid-cols-3 m-auto gap-8 w-[600px] h-[222px]">
                             {menu.map((x) => (
