+++ src/pages/SellingPage.jsx
@@ -4,7 +4,7 @@ import { useDispatch, useSelector } from "react-redux";
 import { fetchProductByUserId, resetSearchProduct } from "../stores/slices/productSlice";
 import { useEffect, useState } from "react";
 import { fetchAllCategory } from "../stores/slices/categorySlice";
-// import Skeleton from "react-loading-skeleton";
+import Skeleton from "react-loading-skeleton";
 import Button from "../components/Button";
 import ListProductContainer from "../features/product/ListProductContainer";
 import Loading from "../components/Loading";
@@ -14,14 +14,14 @@ function SellingPage() {
     const dispatch = useDispatch();
     const { authUserData } = useSelector((state) => state.auth);
     const { productByUserId, loading } = useSelector((state) => state.product);
-    // const [skeleton, setSkeleton] = useState(false);
+    const [skeleton, setSkeleton] = useState(false);
 
-    // useEffect(() => {
-    //     const id = setTimeout(() => {
-    //         setSkeleton(true);
-    //     }, 1200);
-    //     return () => clearTimeout(id);
-    // }, []);
+    useEffect(() => {
+        const id = setTimeout(() => {
+            setSkeleton(true);
+        }, 1200);
+        return () => clearTimeout(id);
+    }, []);
 
     useEffect(() => {
         dispatch(fetchProductByUserId(authUserData?.id));
@@ -41,11 +41,19 @@ function SellingPage() {
 
                             <div className="flex w-full">
                                 <div className="py-3 px-6 w-full mt-16">
-                                    <div className="flex items-center mt-2 py-4 bg-white justify-between rounded-lg">
-                                        <div className="text-xl font-semibold pl-5">Your listings</div>
-                                        <div className="pr-4 ">
-                                            <Search nameTagSearch="" placeholder="Search" />
-                                        </div>
+                                    <div className="flex items-center mt-2 py-4 bg-white justify-between rounded-lg px-5">
+                                        {skeleton ? (
+                                            <div className="text-xl font-semibold ">Your listings</div>
+                                        ) : (
+                                            <Skeleton width={150} height={25} />
+                                        )}
+                                        {skeleton ? (
+                                            <div className="">
+                                                <Search nameTagSearch="" placeholder="Search" />
+                                            </div>
+                                        ) : (
+                                            <Skeleton width={250} height={63} />
+                                        )}
                                     </div>
 
                                     {productByUserId?.length === 0 ? (
