+++ src/pages/SubscribePage.jsx
@@ -2,8 +2,6 @@ import { BsFillPersonFill } from "react-icons/bs";
 import Button from "../components/Button";
 import { useDispatch } from "react-redux";
 import { subscribe } from "../stores/slices/paymentSlice";
-import Skeleton from "react-loading-skeleton";
-import { useEffect, useState } from "react";
 
 function SubscribePage() {
     const dispatch = useDispatch();
@@ -11,14 +9,7 @@ function SubscribePage() {
         e.preventDefault();
         dispatch(subscribe(e.target.name)).then((res) => window.location.replace(res.payload.url));
     };
-    const [skeleton, setSkeleton] = useState(false);
 
-    useEffect(() => {
-        const id = setTimeout(() => {
-            setSkeleton(true);
-        }, 2000);
-        return () => clearTimeout(id);
-    }, []);
     return (
         <div className="flex flex-col items-center justify-center">
             <div className="flex flex-col justify-center items-center h-full mt-16">
@@ -33,32 +24,19 @@ function SubscribePage() {
                             <div className="pr-3 text-[2rem]">
                                 <BsFillPersonFill />
                             </div>
-                            {skeleton ? (
-                                <div className="text-2xl">Individual</div>
-                            ) : (
-                                <Skeleton containerClassName="flex-1" height={40} width={250} />
-                            )}
+
+                            <div className="text-2xl">Individual</div>
                         </div>
                         <div className="px-8 py-4">
-                            {skeleton ? (
-                                <div className="text-xl ">Monthly</div>
-                            ) : (
-                                <Skeleton containerClassName="flex-1" height={20} width={80} />
-                            )}
-                            {skeleton ? (
-                                <div className="font-semibold text-2xl py-2">THB 159.00/month</div>
-                            ) : (
-                                <Skeleton containerClassName="flex-1" height={30} width={200} />
-                            )}
-                            {skeleton ? (
-                                <div className="">
-                                    You will be charged the monthly subscription fee starting on the first billing date.
-                                    The subscription will not automatically renew each month. If you wish to continue,
-                                    please subscribe again at your convenience.
-                                </div>
-                            ) : (
-                                <Skeleton containerClassName="flex-1" count={5} height={15} />
-                            )}
+                            <div className="text-xl ">Monthly</div>
+
+                            <div className="font-semibold text-2xl py-2">THB 159.00/month</div>
+
+                            <div className="">
+                                You will be charged the monthly subscription fee starting on the first billing date. The
+                                subscription will not automatically renew each month. If you wish to continue, please
+                                subscribe again at your convenience.
+                            </div>
                         </div>
                         <div className="flex justify-center pb-4">
                             <div className="flex justify-center w-1/2">
@@ -73,33 +51,18 @@ function SubscribePage() {
                             <div className="pr-3 text-[2rem]">
                                 <BsFillPersonFill />
                             </div>
-                            {skeleton ? (
-                                <div className="text-2xl">Individual</div>
-                            ) : (
-                                <Skeleton containerClassName="flex-1" height={40} width={250} />
-                            )}
+
+                            <div className="text-2xl">Individual</div>
                         </div>
 
                         <div className="px-8 py-4">
-                            {skeleton ? (
-                                <div className="text-xl ">Annual</div>
-                            ) : (
-                                <Skeleton containerClassName="flex-1" height={20} width={80} />
-                            )}
-                            {skeleton ? (
-                                <div className="font-semibold text-2xl py-2">THB 1,590.00/year</div>
-                            ) : (
-                                <Skeleton containerClassName="flex-1" height={30} width={200} />
-                            )}
-                            {skeleton ? (
-                                <div className="">
-                                    You will be charged the monthly subscription fee starting on the first billing date.
-                                    The subscription will not automatically renew each month. If you wish to continue,
-                                    please subscribe again at your convenience.
-                                </div>
-                            ) : (
-                                <Skeleton containerClassName="flex-1" count={5} height={15} />
-                            )}
+                            <div className="text-xl ">Annual</div>
+                            <div className="font-semibold text-2xl py-2">THB 1,590.00/year</div>
+                            <div className="">
+                                You will be charged the monthly subscription fee starting on the first billing date. The
+                                subscription will not automatically renew each month. If you wish to continue, please
+                                subscribe again at your convenience.
+                            </div>
                         </div>
                         <div className="flex justify-center pb-4">
                             <div className="flex justify-center w-1/2">
